new: ParseXML: refactor parsing protocolSupportEnumeration
into a standalone function to make it reusable.

Use it from both parseIDPSSODescriptor and parseSPSSODescriptor

Keep handling of shibboleth10 (generate error) in parseSPSSODescriptor

Note that the results of parsing protocolSupportEnumeration are specific
to a role descriptor, not the whole EntityDescriptor - so should not be stored
as attributes at the top level.