new: allow removing Editors also from showEntity page
The Annual Confirmation page already has UI to remove Editors of an Entity
(users with admin access).

Expose the same functionality also on the showEntity page - it would otherwise
not be intuitive to go to the Annual Confirmation page when looking for a way
to remove admin access for a user.

Show the icons only when the user is either global admin (userLevel>19),
or is an admin of this entity.

Require POST for this action to protect against (some) CSRF attacks:
* in the handler part, require data to come in POST (if at least one parameter
  comes in $_POST, request must have been submitted as a POST).
* in the UI, use a FORM to generate the POST.
  To keep consistent UI (fa-trash icon), submit the form via onClick action
  (a form Button would look quite different).