new: add script for importing ServiceInfo data
Imports from JSON in the same format as produced by the Services API endpoint.

Expects metadata for the SP is already in the metadata tool and allows
to import the service URL record to migrate from tooling that stored the service URL.

This covers the gap that the service URL is not stored in the metadata
and thus does not get imported with the metadata.

Any services that were included in the imported data get flagged as `enabled` for further export.