fix: list scopes by English names of organisations
For deployments without Swedish support, the list at Members->Scopes would show as empty.

As SWAMID rules 5.1.4 and 5.1.5 say that English MUST be provided and Swedish SHOULD be provided,
it is also safer to use English - organizations without a Swedish name would not get shown.

This will turn the list of Scopes to show English names - but as organizations
are also listed on the Members->Organizations list primarily by English name,
this should be presumably acceptable.